% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabnet_index.R
\name{build_tabnet_matrix}
\alias{build_tabnet_matrix}
\title{Build TABNET indicators matrix}
\usage{
build_tabnet_matrix(index, link_start, link_end, name_end, n)
}
\arguments{
\item{index}{HTML scraped object.}

\item{link_start}{List containing start indexes of links, in each item in index object.}

\item{link_end}{List containing end indexes of links, in each item in index object.}

\item{name_end}{List containing end indexes of indicator names, in each item in index object.}

\item{n}{Number of indicators present in each item of index object.}
}
\value{
Data frame with indicators names and links found in objects provided.
}
\description{
Collects indicators links and names from html scraped object.
}
\keyword{tabnet}
