% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabnet_df.R
\name{fetch_all}
\alias{fetch_all}
\title{Retrieve All Indicators}
\usage{
fetch_all(region = "Município",
  url = "http://portal.saude.sp.gov.br/links/matriz", timeout = 1)
}
\arguments{
\item{region}{String specifying desired regionalization. Defaults to "Município"}

\item{url}{TABNET url, defaults to Sao Paulo TABNET.}

\item{timeout}{Time for connection to expire, in seconds.}
}
\value{
Regionalized data frame from selected indicator
}
\description{
Retrieve a gathered data frame with ll indicators from a TABNET link, given a regionalization.
}
\examples{
\dontrun{df <- fetch_all(region = "Região de Saúde")}
}
\keyword{tabnet}
