% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabnet_obj.R
\name{getPostRequestBody}
\alias{getPostRequestBody}
\title{Get POST Request Body}
\usage{
getPostRequestBody(lines, indicators, years)
}
\arguments{
\item{lines}{The lines to be used.}

\item{indicators}{The requested indicators.}

\item{years}{The period of time requested.}
}
\value{
A string properly formatted and encoded to be sent as body of the POST request to TABNET.
}
\description{
Form the raw post request to be uploaded to TABNET.
}
\keyword{tabnet}
