% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabnet_obj.R
\name{getPostURL}
\alias{getPostURL}
\title{Obtain POST url}
\usage{
getPostURL(indicator_url)
}
\arguments{
\item{indicator_url}{The base url of the TABNET indicator.}
}
\value{
A string representing the POST url of the indicator.
}
\description{
Obtain the url to which the POST request should be made, from the base indicator url.
}
\keyword{tabnet}
