% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabnet_df.R
\name{indicator_df}
\alias{indicator_df}
\title{Retrieve Indicator data from its Index}
\usage{
indicator_df(indicator_index, region = NULL, subindicator = NULL,
  years = NULL, onlyMostRecent = FALSE,
  url = "http://portal.saude.sp.gov.br/links/matriz", timeout = 1)
}
\arguments{
\item{indicator_index}{Indicator index number, @seealso [tabnet_index()].}

\item{region}{String describing how should data be regionalized, or an integer indicating region index, @seealso [view_indicator()]}

\item{subindicator}{Among the indicator, which subindicator should be used? Defaults to the last index.}

\item{years}{Specific years desired or their indexes. Defaults to all years available.}

\item{onlyMostRecent}{Boolean. If true, will retrieve only data from the most recent year informed.}

\item{url}{TABNET url, defaults to Sao Paulo TABNET.}

\item{timeout}{Time for connection to expire, in seconds.}
}
\value{
Regionalized data frame from selected indicator
}
\description{
Retrieve a dataframe with indicator data from its index, per region, in certain year period.
}
\examples{
\dontrun{indicator_df(21, region = "DRS", subindicator = 1, years = c(2007, 2008))}
\dontrun{indicator_df(4, region = "RRAS", onlyMostRecent = TRUE)}
}
\seealso{
[tabnet_df_retrieval()]

[tabnet_csv_retrieval()]
}
\keyword{tabnet}
