% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabnet_index.R
\name{indicator_list}
\alias{indicator_list}
\title{Indicator List}
\usage{
indicator_list(url = "http://portal.saude.sp.gov.br/links/matriz")
}
\arguments{
\item{url}{TABNET Url.}
}
\value{
If provided a valid link, a vector of available TABNET indicators.
}
\description{
Retrieve a vector of available indicators from link provided.
}
\keyword{tabnet}
