% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabnet_index.R
\name{indicator_search}
\alias{indicator_search}
\title{Search Indicators' Names from a Keyword}
\usage{
indicator_search(keywords,
  url = "http://portal.saude.sp.gov.br/links/matriz")
}
\arguments{
\item{url}{TABNET Url.}

\item{keyword}{Keyword to be searched.}
}
\value{
Subset of vector of indicator names that match specified keyword.
}
\description{
Retrieve a subset of available indicators from list that match the specified keyword.
}
\examples{
\dontrun{indicator_search("sífilis")}
}
\keyword{tabnet}
