% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabnet_df.R
\name{is.tabnet_obj}
\alias{is.tabnet_obj}
\title{TABNET Object Check}
\usage{
is.tabnet_obj(obj)
}
\arguments{
\item{obj}{Object to be checked}
}
\value{
Boolean, indicating whether object passed is a valid TABNET object
}
\description{
Check if a certain object has the appropriate TABNET object parameters.
}
\seealso{
[make_tabnet_obj()]
}
\keyword{tabnet}
