% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabnet_df.R
\name{isRegionAvailable}
\alias{isRegionAvailable}
\title{Regionalization Check}
\usage{
isRegionAvailable(obj, region = "Município")
}
\arguments{
\item{obj}{TABNET object to be checked. @seealso [make_tabnet_obj()]}

\item{region}{String indicating regionalization to be tested.}
}
\value{
Boolean, indicating whether the TABNET object has desired regionalization among its avaiable ones.
}
\description{
Check if a certain indicator supports desired regionalization.
}
\keyword{tabnet}
