% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_handler.R
\name{safe_POST}
\alias{safe_POST}
\title{Safe POST Operation}
\usage{
safe_POST(url, body, encode = "raw", timeout = 1, num_attempts = 5)
}
\arguments{
\item{url}{POST request url.}

\item{timeout}{Time (seconds) to wait in each attempt.}

\item{num_attempts}{Maximum number of attempts to be made.}
}
\value{
Response instance containing possible error message and the http response content.
}
\description{
Perform several attempts of an HTTP POST operation.
}
\keyword{tabnet}
