% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabnet_encode.R
\name{tabnet_breakline_fix}
\alias{tabnet_breakline_fix}
\title{TABNET Breakline fix}
\usage{
tabnet_breakline_fix(lines)
}
\arguments{
\item{lines}{Lines to be fixed.}
}
\value{
The same line, without break characters.
}
\description{
Fix the line breaks of indicators.
}
\keyword{tabnet}
