% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabnet_encode.R
\name{tabnet_concatFormat}
\alias{tabnet_concatFormat}
\title{Concatenate TABNET Files}
\usage{
tabnet_concatFormat(elements, sep = "Arquivos")
}
\arguments{
\item{elements}{The files to be concatenated.}

\item{sep}{The separator between the file names.}
}
\value{
A string concatenating all elements provided using the separator.
}
\description{
When asking for more than one file, the names must be concatenated in a certain pattern.
}
\keyword{tabnet}
