% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabnet_csv_retrieval.R
\name{tabnet_csv_retrieval}
\alias{tabnet_csv_retrieval}
\title{Retrieve TABNET data in .csv format}
\usage{
tabnet_csv_retrieval(post_url, body, colname = "Região",
  ind_name = "Valor", skip = 3, stopDelta = 7, timeout = 4)
}
\arguments{
\item{post_url}{Indicator POST url.}

\item{body}{Request body containing required information.}

\item{colname}{The specified region name, like city.}

\item{skip}{Numeric interval from which the .csv will start to be read.}

\item{stopDelta}{Numeric interval from which the .csv reading will stop.}

\item{timeout}{Maximum time interval POST request should take.}
}
\value{
Data frame containing data from the indicator specified by the url.
}
\description{
Retrieve the indicator data from its url and parameters.
}
\keyword{tabnet}
