% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabnet_df.R
\name{tabnet_df}
\alias{tabnet_df}
\title{Retrieve TABNET data in dataframe form}
\usage{
tabnet_df(indicator_index, region = "Município", subindicator = NULL,
  years = NULL, onlyMostRecent = FALSE)
}
\arguments{
\item{indicator_index}{Indicator index number, @seealso [tabnet_index()].}

\item{region}{String describing how should data be regionalized.}

\item{subindicator}{Among the indicator, which subindicator should be used? Defaults to the last index.}

\item{years}{Retrieve data from one or more specific years?}

\item{onlyMostRecent}{Boolean. If true, will retrieve only data from the most recent year informed.}
}
\value{
Regionalized data frame from selected indicator
}
\description{
Retrieve a dataframe with indicator data, per region, per year.
}
\seealso{
[tabnet_csv_retrieval()]
}
\keyword{tabnet}
