% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabnet_df.R
\name{tabnet_df_retrieval}
\alias{tabnet_df_retrieval}
\title{TABNET dataframe retrieval}
\usage{
tabnet_df_retrieval(obj, line_index = 2, ind_index = NULL,
  years_index = NULL, onlyMostRecent = FALSE, timeout = 4)
}
\arguments{
\item{obj}{TABNET Object}

\item{line_index}{Chosen line index, defaults to 2 - city}

\item{ind_index}{Chosen subindicator index, defaults to last one}

\item{years_index}{Chosen years/years, defaults to most recent available}

\item{onlyMostRecent}{Boolean. Should only data from most recent year informed be returned?}
}
\value{
Corresponding data frame
}
\description{
Given a TABNET Object and certain parameters, automatically retrieve corresponding data frame
}
\keyword{tabnet}
