% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabnet_encode.R
\name{tabnet_hex_escape}
\alias{tabnet_hex_escape}
\title{Tabnet Hexadecimal Escape Function}
\usage{
tabnet_hex_escape(indicators)
}
\arguments{
\item{indicators}{Indicator strings.}
}
\value{
String with adapted escaped characters.
}
\description{
Adapt parts of the escaped string so it is uploadable to TABNET.
}
\keyword{tabnet}
