% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabnet_index.R
\name{tabnet_index}
\alias{tabnet_index}
\title{Indicators Matrix}
\usage{
tabnet_index(url = "http://portal.saude.sp.gov.br/links/matriz")
}
\arguments{
\item{url}{TABNET Url.}
}
\value{
If provided a valid link, a data frame with names and links of each indicator will be returned.
}
\description{
Retrieve a matrix from TABNET of available indicators and their web links.
}
\keyword{tabnet}
