% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabnet_encode.R
\name{tabnet_indicator_catch}
\alias{tabnet_indicator_catch}
\title{Indicator Check}
\usage{
tabnet_indicator_catch(indicators)
}
\arguments{
\item{indicators}{Indicator strings.}
}
\value{
Return clean indicator strings.
}
\description{
Clean an indicator string.
}
\keyword{tabnet}
