% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabnet_map.R
\name{tabnet_map}
\alias{tabnet_map}
\title{Generate Map plot from indicator data}
\usage{
tabnet_map(indicator_index, region = "Município", subindicator = NULL,
  years = NULL, label = FALSE, nBins = NULL, palette = "Purples",
  title = NULL)
}
\arguments{
\item{indicator_index}{Indicator index number, @seealso [tabnet_index()].}

\item{region}{String describing how should data be regionalized.}

\item{subindicator}{Among the indicator, which subindicator should be used? Defaults to the last index.}

\item{years}{Retrieve data from one or more specific years?}

\item{label}{Boolean. Should the regions be labeled? Defaults to false.}

\item{nBins}{Number of bins to discretize desired indicator value.}

\item{palette}{RColorBrewer palette.}

\item{title}{Do you want to provide a specific title? Otherwise, function will generate one itself.}
}
\value{
ggplot2 object
}
\description{
Generate a ggplot2 map from indicator data of a specific year.
}
\keyword{tabnet}
