% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabnet_map2.R
\name{tabnet_map2}
\alias{tabnet_map2}
\title{Generate Map plot from indicator data, with North indicator and scale}
\usage{
tabnet_map2(
  indicator_index,
  region = "Município",
  subindicator = NULL,
  years = NULL,
  label = FALSE,
  nBins = NULL,
  palette = "Purples",
  title = NULL,
  scalebar = TRUE,
  north_arrow = TRUE,
  scale_width = 0.3
)
}
\arguments{
\item{indicator_index}{Indicator index number, @seealso [tabnet_index()].}

\item{region}{String describing how should data be regionalized.}

\item{subindicator}{Among the indicator, which subindicator should be used? Defaults to the last index.}

\item{years}{Retrieve data from one or more specific years?}

\item{label}{Boolean. Should the regions be labeled? Defaults to false.}

\item{nBins}{Number of bins to discretize desired indicator value.}

\item{palette}{RColorBrewer palette.}

\item{title}{Do you want to provide a specific title? Otherwise, function will generate one itself.}

\item{scalebar}{Boolean, indicating whether a scale bar should be present on the map. Defaults to true.}

\item{north_arrow}{Boolean, indicating whether a North arrow should be added on the map. Defaults to true.}

\item{scale_width}{ggspatial parameter, indicating scale bar width.}
}
\value{
ggplot2 object
}
\description{
Other version of [tabnet_map()], with North indicators and scale bar. Requires ggspatial package.
}
\keyword{tabnet}
