% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabnet_encode.R
\name{tabnet_special_character_check}
\alias{tabnet_special_character_check}
\title{TABNET Special Character Check}
\usage{
tabnet_special_character_check(escaped_string)
}
\arguments{
\item{escaped_string}{String already escaped.}
}
\value{
The same string, with modifications if special characters are present.
}
\description{
Some characters need special attention when generating an encoded string to TABNET.
}
\keyword{tabnet}
