% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabnet_obj.R
\name{view_indicator}
\alias{view_indicator}
\title{View Indicator Information}
\usage{
view_indicator(
  indicator_index,
  url = "http://portal.saude.sp.gov.br/links/matriz",
  timeout = 1
)
}
\arguments{
\item{indicator_index}{Index of desired indicator, according to Indicator List.}

\item{url}{URL for indicator retrieval, defaults to SP TABNET.}

\item{timeout}{Timeout period, in seconds.}
}
\value{
An object containing information about the indicator, its parameters, available years, regions and subindicators.
}
\description{
View available indicator regions, years and subindicators.
}
\examples{
\dontrun{view_indicator(21)}
}
\keyword{tabnet}
