% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabnet_encode.R
\name{W1252.encode}
\alias{W1252.encode}
\title{W1252 Encode}
\usage{
W1252.encode(decoded)
}
\arguments{
\item{decoded}{The regular string, before encoding.}
}
\value{
An encoded string according to TABNET's W1252 pattern.
}
\description{
Process the string to the W1252 encoding, used by the TABNET System.
}
\keyword{tabnet}
